// src/models/quizzr.ts
export class Quizzr{
    
   title:any;
   description: any;
   quiz_type: any;
   start_date: any;
   end_date :any;
   duration_mins: number;
   attempts_allowed :number;
   randomized:boolean;
   question_number:number;
   question: any;
   sequence_number: number;
   answer_type:any;
   correct_answer_count: number;
   explanation: any;
   answer: any;
   is_right_answer:boolean;
   qid: number;
   user_id:number;
   start_time:string;
   submit_time:string;
   score: number;
   attempt_id: number;

}